# docs/api-contract.yaml
openapi: "3.1.0"
info:
  title: Urban Sustainability Assessment System — API
  version: "1.0.0"
  description: |
    REST API for S-50 HackNova'26. Provides sustainability scores, indicator data,
    recommendations, and policy simulation for 10 Chennai wards across 5 SDG-11 pillars.

    ## ⚠️  CONTRACT FROZEN — Day 3
    **Do NOT change field names, types, or endpoint paths without unanimous team sign-off.**
    Any change from here needs approval from whoever is building against it (M3 / M4).
    Breaking this contract mid-build is how the schedule slips.

  contact:
    name: Team USAS
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://usas-api.onrender.com
    description: Staging (Render)
...
paths:
...
  /api/v1/wards/{ward_id}:
    get:
      operationId: get_ward
      tags: [wards]
      summary: Get a single ward
      parameters:
        - name: ward_id
          in: path
          required: true
          schema:
            type: string
            example: ward-001
      responses:
        "200":
          description: Ward detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WardOut"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Indicators ────────────────────────────────────────────────────────────
...
components:
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Ward 'ward-099' not found."
...
  schemas:
...
    WardOut:
      type: object
      required: [id, name, city, state, lat, lon, population, area_km2]
      properties:
        id:          { type: string, example: ward-001 }
        name:        { type: string, example: Anna Nagar }
        city:        { type: string, example: Chennai }
        state:       { type: string, example: Tamil Nadu }
        lat:         { type: number, format: float, example: 13.0850 }
        lon:         { type: number, format: float, example: 80.2101 }
        population:  { type: integer, example: 125000 }
        area_km2:    { type: number, format: float, example: 8.5 }
...